+++ src/Runtime/Matcher/MatcherFactory.php
@@ -3,17 +3,8 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Path\Runtime\Matcher;
 
-use Remorhaz\JSON\Path\Runtime\ValueFetcherInterface;
-
 final class MatcherFactory implements MatcherFactoryInterface
 {
-
-    private $valueFetcher;
-
-    public function __construct(ValueFetcherInterface $valueFetcher)
-    {
-        $this->valueFetcher = $valueFetcher;
-    }
 
     public function matchAnyChild(): ChildMatcherInterface
     {
@@ -32,6 +23,6 @@ final class MatcherFactory implements MatcherFactoryInterface
 
     public function matchElementSlice(?int $start, ?int $end, ?int $step): ChildMatcherInterface
     {
-        return new SliceElementMatcher($this->valueFetcher, $start, $end, $step);
+        return new SliceElementMatcher($start, $end, $step);
     }
 }