+++ src/Runtime/ValueFetcher.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 namespace Remorhaz\JSON\Path\Runtime;
 
 use Remorhaz\JSON\Data\Value\StructValueInterface;
-use function iterator_count;
-use Remorhaz\JSON\Data\Value\ArrayValueInterface;
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
 use Remorhaz\JSON\Data\Value\ScalarValueInterface;
 
@@ -26,7 +24,7 @@ final class ValueFetcher implements ValueFetcherInterface
         }
 
         if ($value instanceof StructValueInterface) {
-            return $this->fetchChildren($value, $matcher);
+            return $this->fetchStructChildren($value, $matcher);
         }
 
         throw new Exception\UnexpectedNodeValueFetchedException($value);
@@ -41,13 +39,13 @@ final class ValueFetcher implements ValueFetcherInterface
         }
 
         if ($value instanceof StructValueInterface) {
-            return $this->fetchDeepChildren($value, $matcher);
+            return $this->fetchStructDeepChildren($value, $matcher);
         }
 
         throw new Exception\UnexpectedNodeValueFetchedException($value);
     }
 
-    private function fetchChildren(NodeValueInterface $value, Matcher\ChildMatcherInterface $matcher): array
+    private function fetchStructChildren(NodeValueInterface $value, Matcher\ChildMatcherInterface $matcher): array
     {
         if (!$value instanceof StructValueInterface) {
             // TODO: extract correct argument type?
@@ -63,7 +61,7 @@ final class ValueFetcher implements ValueFetcherInterface
         return $results;
     }
 
-    private function fetchDeepChildren(NodeValueInterface $value, Matcher\ChildMatcherInterface $matcher): array
+    private function fetchStructDeepChildren(NodeValueInterface $value, Matcher\ChildMatcherInterface $matcher): array
     {
         if (!$value instanceof StructValueInterface) {
             // TODO: extract correct argument type?
@@ -82,12 +80,5 @@ final class ValueFetcher implements ValueFetcherInterface
         }
 
         return $results;
-    }
-
-    public function fetchArrayLength(NodeValueInterface $value): ?int
-    {
-        return $value instanceof ArrayValueInterface
-            ? iterator_count($value->createChildIterator())
-            : null;
     }
 }